Move some startup methods to StartupHelpers
